requirements.txt
google-genai
matplotlib
python-dotenv
